Added lighting and enemies
Enemies follow the player when he is in their room. The player attacks can kill the enemies. I also added 2D lighting to the scene. The player sprite now changes colour when he selects different armours, and an icon pops up to display the artifact chosen. Their is also now an end game screen that occurs when you hit the stairs at the end of the level